feat: Add CRUD for Orders ,Products and Payments

+++ src/main/java/com/pe/upc/smartgardenbackend/ecommerce/domain/model/commands/DeleteOrderCommand.java
@@ -1,4 +1,3 @@
 package com.pe.upc.smartgardenbackend.ecommerce.domain.model.commands;
 
-public record DeleteOrderCommand(int id) {
-}
+public record DeleteOrderCommand(Long id) {}
